refactor: split models.rs into focused domain submodules [OPE-484]
Break 494-line monolithic models.rs into four focused submodules:
- session.rs: Session, Message, MessageQueue, AgentMessage models
- work.rs: WorkItem and OrchestrationRun models
- monitoring.rs: AlertRule, AlertHistory, EventHistory models
- infrastructure.rs: Schedule, Plugin, Trigger, ApiKey models

All 259 existing tests pass. No public API changes — mod.rs
re-exports everything via glob imports.
