+++ crates/opengoose-persistence/src/models/mod.rs
@@ -1,0 +1,12 @@
+mod infrastructure;
+mod monitoring;
+mod session;
+mod work;
+
+pub use infrastructure::*;
+pub use monitoring::*;
+pub use session::*;
+pub use work::*;
+
+#[cfg(test)]
+mod tests;
