+++ crates/opengoose-persistence/src/models/work.rs
@@ -1,0 +1,64 @@
+use diesel::prelude::*;
+
+use crate::schema::*;
+
+// ── Work Items ──
+
+#[derive(Queryable, Selectable)]
+#[diesel(table_name = work_items)]
+pub struct WorkItemRow {
+    pub id: i32,
+    pub session_key: String,
+    pub team_run_id: String,
+    pub parent_id: Option<i32>,
+    pub title: String,
+    pub description: Option<String>,
+    pub status: String,
+    pub assigned_to: Option<String>,
+    pub workflow_step: Option<i32>,
+    pub input: Option<String>,
+    pub output: Option<String>,
+    pub error: Option<String>,
+    pub created_at: String,
+    pub updated_at: String,
+}
+
+#[derive(Insertable)]
+#[diesel(table_name = work_items)]
+pub struct NewWorkItem<'a> {
+    pub session_key: &'a str,
+    pub team_run_id: &'a str,
+    pub parent_id: Option<i32>,
+    pub title: &'a str,
+}
+
+// ── Orchestration Runs ──
+
+#[derive(Queryable, Selectable)]
+#[diesel(table_name = orchestration_runs)]
+pub struct OrchestrationRunRow {
+    #[allow(dead_code)]
+    pub id: i32,
+    pub team_run_id: String,
+    pub session_key: String,
+    pub team_name: String,
+    pub workflow: String,
+    pub input: String,
+    pub status: String,
+    pub current_step: i32,
+    pub total_steps: i32,
+    pub result: Option<String>,
+    pub created_at: String,
+    pub updated_at: String,
+}
+
+#[derive(Insertable)]
+#[diesel(table_name = orchestration_runs)]
+pub struct NewOrchestrationRun<'a> {
+    pub team_run_id: &'a str,
+    pub session_key: &'a str,
+    pub team_name: &'a str,
+    pub workflow: &'a str,
+    pub input: &'a str,
+    pub total_steps: i32,
+}
